+++ glean/src/core/glean.ts
@@ -19,6 +19,7 @@ import MetricsDatabase from "./metrics/database.js";
 import EventsDatabase from "./metrics/events_database/index.js";
 import PingsDatabase from "./pings/database.js";
 import ErrorManager from "./error/index.js";
+import GleanMetrics from "./glean_metrics.js";
 
 const LOG_TAG = "core.Glean";
 
@@ -317,6 +318,12 @@ namespace Glean {
       onUploadEnabled();
       initializeCoreMetrics(config?.migrateFromLegacyStorage);
 
+      // Record a page load event if the client has auto page-load events enabled.
+      if (config?.enableAutoPageLoadEvents) {
+        // This function call has no parameters because auto-instrumentation
+        // means there are no overrides.
+        GleanMetrics.pageLoad();
+      }
     } else {
       // If upload is disabled, and we've never run before, only set the
       // client_id to KNOWN_CLIENT_ID, but do not send a deletion request