+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 [#1808](https://github.com/mozilla/glean.js/pull/1808): **BREAKING CHANGE**: Make glean.js fully synchronous.
 
+* [#1835](https://github.com/mozilla/glean.js/pull/1835): Automatic instrumentation of page load events for simple web properties.
+
 # v3.0.0 (2023-11-16)
 
 [Full changelog](https://github.com/mozilla/glean.js/compare/v3.0.0-pre.1...v3.0.0)